38: code refactor, abstract tabs and floating button into separate files
