+++ app/src/main/java/com/example/conquest/components/MyFloatingActionButton.kt
@@ -1,0 +1,30 @@
+package com.example.conquest.components
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun MyFab(
+    onClick: () -> Unit, modifier: Modifier = Modifier
+) {
+    FloatingActionButton(
+        onClick = onClick,
+        containerColor = MaterialTheme.colorScheme.tertiary,
+        contentColor = MaterialTheme.colorScheme.primary,
+        shape = RoundedCornerShape(50),
+        modifier = modifier
+            .padding(bottom = 16.dp)
+            .statusBarsPadding()
+    ) {
+        Icon(Icons.Default.Add, contentDescription = "Add")
+    }
+}
