+++ app/src/main/java/com/example/conquest/screens/CosplayElementsTab.kt
@@ -1,0 +1,16 @@
+package com.example.conquest.screens
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+
+@Composable
+fun CosplayElementsTab() {
+    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+        Text("Cosplay Elements Content")
+    }
+}
